requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.1.8
clip==1.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.19.1
fsspec==2025.10.0
ftfy==6.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.13.0.90
packaging==26.0
pillow==11.3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.11.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
python-dotenv==1.2.1
python-socketio==5.16.1
python-engineio==4.13.1
redis[asyncio]>=5.0.0
simple-websocket==1.1.0
bidict==0.23.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2026.1.15
rich==14.3.1
rich-toolkit==0.17.1
starlette==0.49.3
sympy==1.14.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.5.0
websockets==15.0.1
